Move stuff to new class and order commands
