+++ newspack-custom-content-migrator.php
@@ -60,7 +60,6 @@ PluginSetup::register_migrators(
 		Command\General\PrelaunchSiteQAMigrator::class,
 		Command\General\VillageMediaCMSMigrator::class,
 		Command\General\MetroMigrator::class,
-		Command\General\PloneCMSMigrator::class,
 		Command\General\ProfilePress::class,
 		Command\General\TownNewsMigrator::class,
 		Command\General\UsersMigrator::class,
@@ -74,7 +73,7 @@ PluginSetup::register_migrators(
 		Command\PublisherSpecific\BethesdaMagMigrator::class,
 		Command\PublisherSpecific\SearchLightNMMigrator::class,
 		Command\PublisherSpecific\CalMattersMigrator::class,
-		Command\PublisherSpecific\HighCountryNewsMigrator::class,
+		Command\PublisherSpecific\HighCountryNewsMigrator2::class,
 		Command\PublisherSpecific\NewsroomCoNzMigrator::class,
 		Command\PublisherSpecific\MassterlistMigrator::class,
 		Command\PublisherSpecific\MauiTimesMigrator::class,
